Add files via upload
 R Shiny application is based off of a dataset of Airbnb listings in New York City for 2019. 
(Data Link :https://www.kaggle.com/dgomonov/new-york-city-airbnb-open-data#AB_NYC_2019.csv)
our application launches the user will be able to see a menu bar at the top in the form of tabs where they can easily navigate between the different features.  There is also a background of New York City for aesthetics and to fit the theme of our data.
In the first tab of our application we have subsetted the dataset to include the columns that users would likely be most interested in when searching for an Airbnb. 
In the second tab we have included a tab with information on availability.  The user can select a borough and a price range, and the app will show them a histogram of how many days out of the year Airbnb’s with those parameters are available.  
In the third tab we gave the user a highly interactive way of exploring Airbnb prices in New York City. 
In the final tab we used a new packaged called leaflet to plot the Airbnb locations on a map from their longitude and latitude coordinates.